Refactor component imports to use unified Box component path
- Updated imports in BlogError, BlogPost, ProjectCard, ResourceCard, NewsletterCTA, and VideoCard components to reference the new Box component location in the ui directory.
- Deleted the outdated Box component from the layout directory along with its associated test file to streamline the codebase and improve organization.
